+++ packages/terra-collapsible-menu-view/tests/nightwatch/DefaultCollapsibleMenuView.jsx
@@ -17,7 +17,7 @@ class CollapsibleMenuViewDemo extends React.Component {
     this.state = {
       toggle1Selection: false,
       toggle2Selection: false,
-      displayType: 0,
+      displayType: 'tableView',
     };
   }
 
@@ -29,8 +29,8 @@ class CollapsibleMenuViewDemo extends React.Component {
     this.setState({ toggle2Selection: isSelected });
   }
 
-  handleDisplayTypeChange(event, selectedIndex) {
-    this.setState({ displayType: selectedIndex });
+  handleDisplayTypeChange(event, selectedKey) {
+    this.setState({ displayType: selectedKey });
   }
 
   render() {
@@ -74,7 +74,7 @@ class CollapsibleMenuViewDemo extends React.Component {
           ]}
         />
         <CollapsibleMenuView.Divider key="Divider2" />
-        <CollapsibleMenuView.ItemGroup key="ViewTypeSelection" className="ViewTypeGroup" isSelectable onChange={this.handleDisplayTypeChange}>
+        <CollapsibleMenuView.ItemGroup key="ViewTypeSelection" className="ViewTypeGroup" selectedKeys={[this.state.displayType]} onChange={this.handleDisplayTypeChange}>
           <CollapsibleMenuView.Item
             icon={<IconTable />}
             text="Table View"
@@ -82,7 +82,7 @@ class CollapsibleMenuViewDemo extends React.Component {
             key="tableView"
             isIconOnly
             shouldCloseOnClick={false}
-            isSelected={this.state.displayType === 0}
+            isSelected={this.state.displayType === 'tableView'}
           />
           <CollapsibleMenuView.Item
             icon={<IconFlowsheet />}
@@ -91,7 +91,7 @@ class CollapsibleMenuViewDemo extends React.Component {
             key="expandedView"
             isIconOnly
             shouldCloseOnClick={false}
-            isSelected={this.state.displayType === 1}
+            isSelected={this.state.displayType === 'expandedView'}
           />
           <CollapsibleMenuView.Item
             icon={<IconVisualization />}
@@ -100,7 +100,7 @@ class CollapsibleMenuViewDemo extends React.Component {
             key="trendingView"
             isIconOnly
             shouldCloseOnClick={false}
-            isSelected={this.state.displayType === 2}
+            isSelected={this.state.displayType === 'trendingView'}
           />
         </CollapsibleMenuView.ItemGroup>
         <CollapsibleMenuView.Divider key="Divider3" />